Move email validation to a shared concern
